https://github.com/jackdewinter/pymarkdown/issues/600 - refactored html helper.
