+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14479,
-        "totalCovered": 14479
+        "totalMeasured": 14501,
+        "totalCovered": 14501
     }
 }
