+++ pymarkdown/leaf_blocks/leaf_block_processor.py
@@ -5,7 +5,7 @@ import logging
 from typing import List, Optional
 
 from pymarkdown.container_grab_bag import ContainerGrabBag
-from pymarkdown.html_helper import HtmlHelper
+from pymarkdown.html.html_helper import HtmlHelper
 from pymarkdown.leaf_blocks.atx_leaf_block_processor import AtxLeafBlockProcessor
 from pymarkdown.leaf_blocks.fenced_leaf_block_processor import FencedLeafBlockProcessor
 from pymarkdown.leaf_blocks.leaf_block_helper import LeafBlockHelper
